roachprod: retry hostname validation on different nodes
Previously hostname validation was only attempted on the
first node. If the first node reports a mismatch in names,
the entire cluster should mismatch as well.

However, it's possible that the first node fails due to
a transient error. In this case, we want to try other nodes
as a VM being down does not mean the hostnames are wrong.

This is an important distinction for artifacts collection.
If a test failed because the a VM went down, we still want
to collect logs from the other nodes.